+++ jina/checker.py
@@ -51,7 +51,7 @@ class NetworkChecker:
     """Check if a BasePod is running or not """
 
     def __init__(self, args: 'argparse.Namespace'):
-        from jina.peapods.pea import send_ctrl_message
+        from jina.peapods.peas import send_ctrl_message
         from jina.logging.profile import TimeContext
         from google.protobuf.json_format import MessageToJson
         import time
