+++ jina/peapods/peas/remote_pea/ssh.py
@@ -1,9 +1,9 @@
 from subprocess import Popen, PIPE
 
-from ..pea import BasePea
-from ... import __ready_msg__, __stop_msg__
-from ...helper import get_non_defaults_args, kwargs2list
-from ...logging import JinaLogger
+from jina.peapods.peas import BasePea
+from jina import __ready_msg__, __stop_msg__
+from jina.helper import get_non_defaults_args, kwargs2list
+from jina.logging import JinaLogger
 
 
 class RemoteSSHPea(BasePea):
@@ -22,7 +22,7 @@ class RemoteSSHPea(BasePea):
 
     @property
     def remote_command(self) -> str:
-        from ..parser import set_pea_parser
+        from ...parser import set_pea_parser
         non_defaults = get_non_defaults_args(self.args, set_pea_parser(), taboo={'host'})
         _args = kwargs2list(non_defaults)
         return f'jina pea {" ".join(_args)}'
@@ -56,7 +56,7 @@ class RemoteSSHPod(RemoteSSHPea):
 
     @property
     def remote_command(self) -> str:
-        from ..parser import set_pod_parser
+        from ...parser import set_pod_parser
         non_defaults = get_non_defaults_args(self.args, set_pod_parser(), taboo={'host'})
         _args = kwargs2list(non_defaults)
         return f'jina pod {" ".join(_args)}'
