+++ tests/unit/peapods/test_remote.py
@@ -4,7 +4,7 @@ from jina.parser import set_gateway_parser, set_pea_parser
 from jina.peapods.pod.gateway_pod import GatewayPod
 
 if False:
-    from jina.peapods.remote import PeaSpawnHelper
+    from jina.peapods.peas.remote_pea import PeaSpawnHelper
 
 
 @pytest.mark.skip
