+++ tests/unit/test_echostream.py
@@ -2,7 +2,7 @@ import logging
 
 from jina.flow import Flow
 from jina.parser import set_pea_parser
-from jina.peapods.pea import BasePea
+from jina.peapods.peas import BasePea
 from jina.peapods.zmq import Zmqlet
 from jina.proto import jina_pb2
 from jina.types.message import Message
