+++ tests/unit/test_is_read_message.py
@@ -1,6 +1,6 @@
 from jina.logging import default_logger
 from jina.parser import set_pea_parser
-from jina.peapods.pea import BasePea
+from jina.peapods.peas import BasePea
 from jina.peapods.zmq import Zmqlet
 from jina.proto import jina_pb2
 from jina.types.message import Message
